SAS Code with manual entered data.
This SAS code snippet is used to create a dataset called "characters" with information about fictional characters. The data is manually entered into the code using the "input" statement, which specifies the variables: name, gender, height, and weight. The "cards" statement is followed by the actual data entries for each character, including their name, gender, height (in centimeters), and weight (in kilograms).